server/pyproject.toml
[project]
name = "fastapi-vuejs-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.2,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "email-validator>=2.1.0",
    "structlog>=24.4.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "mypy>=1.14.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
    "aiosqlite>=0.19.0",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = "."

[[tool.mypy.overrides]]
module = ["jose.*", "passlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
